requirements.txt
streamlit
pandas
plotly
openpyxl
python-pptx
kaleido==0.2.1.post1
